Update to v0.3.2 (#32)
* upgrade: update from cryptogarageinc/v0.3.10

* fix: add exec

+++ wrap_js/__integration_test__/bitcoin-docker-entrypoint.sh
@@ -1,22 +1,20 @@
 #!/bin/bash -u
 
 # while :; do sleep 10; done
-export WORKDIR_ROOT=root
-export WORK_DIR=wallet-test
-export WORKDIR_PATH=/${WORKDIR_ROOT}/${WORK_DIR}
+export WORKDIR_ROOT=workspace
+export WORKDIR_PATH=/workspace
 
-cd /${WORKDIR_ROOT}
-if [ ! -d ${WORK_DIR} ]; then
-  mkdir ${WORK_DIR}
+if [ ! -d ${WORKDIR_PATH} ]; then
+  mkdir ${WORKDIR_PATH}
 fi
 
-cd ${WORKDIR_PATH}
+cd /${WORKDIR_ROOT}
 rm -rf bitcoind_datadir
 
 mkdir bitcoind_datadir
 chmod 777 bitcoind_datadir
 # cp /root/.bitcoin/bitcoin.conf bitcoind_datadir/
-cp ./__tests__/bitcoin.conf bitcoind_datadir/
+cp ./wrap_js/__integration_test__/bitcoin.conf bitcoind_datadir/
 
 # boot daemon
 bitcoind --regtest -datadir=${WORKDIR_PATH}/bitcoind_datadir
@@ -29,11 +27,32 @@ echo "start bitcoin node"
 
 set -e
 
-if [ -d node_modules ]; then
+clrFlg="true"
+if [ $# -gt 0 ]; then
+  echo "param: $1"
+  if [ "$1" = "cache" ]; then
+    clrFlg="false"
+  fi
+fi
+echo "clrFlg: $clrFlg"
+
+if [ "$clrFlg" = "true" ]; then
+  if [ -d node_modules ]; then
+    rm -rf node_modules
+  fi
+  mkdir node_modules
+  chmod 777 node_modules
+fi
+
+node --version
+
+if [ "$clrFlg" = "true" ]; then
+  npm install
+  npm run cmake_release_parallel
+fi
+
+npm run bitcoin_test
+
+if [ "$clrFlg" = "true" ]; then
   rm -rf node_modules
 fi
-mkdir node_modules
-chmod 777 node_modules
-node --version
-npm install && npm bitcoin_test
-rm -rf node_modules
